requirements.txt
alembic==1.8.1
asttokens==2.0.8
backcall==0.2.0
bcrypt==4.0.0
cffi==1.15.1
cli-helpers==2.2.1
click==8.1.3
colorama==0.4.5
configobj==5.0.6
cryptography==37.0.4
decorator==5.1.1
executing==0.10.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
parso==0.8.3
pendulum==2.1.2
pgcli==3.4.1
pgspecial==1.13.1
pickleshare==0.7.5
prompt-toolkit==3.0.30
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
python-dateutil==2.8.2
pytzdata==2020.1
setproctitle==1.3.2
six==1.16.0
SQLAlchemy==1.4.40
sqlparse==0.4.2
stack-data==0.4.0
tabulate==0.8.10
traitlets==5.3.0
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.1
